Fix warning for _YIELD_VALUE

+++ Python/executor_cases.c.h
@@ -9346,7 +9346,7 @@
             assert(INLINE_CACHE_ENTRIES_SEND == INLINE_CACHE_ENTRIES_FOR_ITER);
             #if TIER_ONE && defined(Py_DEBUG)
             if (!PyStackRef_IsNone(frame->f_executable)) {
-                int i = frame->instr_ptr - _PyFrame_GetBytecode(frame);
+                int i = (int)(frame->instr_ptr - _PyFrame_GetBytecode(frame));
                 int opcode = _Py_GetBaseCodeUnit(_PyFrame_GetCode(frame), i).op.code;
                 assert(opcode == SEND || opcode == FOR_ITER);
             }
